Added a NucleaseClass class to record protein/mRNA and converted concentrations to Double values.

+++ impc_prod_tracker/src/main/java/uk/ac/ebi/impc_prod_tracker/data/biology/attempt/crispr_attempt/CrisprAttempt.java
@@ -57,7 +57,7 @@ public class CrisprAttempt extends BaseEntity
 
     private Boolean guidesGeneratedInPlasmid;
 
-    private Float grnaConcentration;
+    private Double grnaConcentration;
 
     @Column(name = "no_g0_where_mutation_detected")
     private Integer noG0WhereMutationDetected;
