+++ impc_prod_tracker/src/main/java/uk/ac/ebi/impc_prod_tracker/data/biology/mutagenesis_donor/MutagenesisDonor.java
@@ -23,7 +23,7 @@ public class MutagenesisDonor extends BaseEntity
     @ManyToOne(targetEntity = CrisprAttempt.class)
     private CrisprAttempt crisprAttempt;
 
-    private Integer concentration;
+    private Double concentration;
 
     private String preparation;
